Logged in user is falsely flagged as guest preventing autoenrolment on login
When you have an active session as a guest user and then login or create a real user account. The autoenrolment on login will not enrol the user because the session still contains the guest user object instead of the logged in user and this logged in user is not passed as argument to the `isguestuser()` method as argument. And the autoenrol therefor skips enrolling guest users.